separate minicpmo data
