+++ verl/workers/actor/dp_actor.py
@@ -29,6 +29,7 @@ from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
 import verl.utils.torch_functional as verl_F
 from verl import DataProto
 from verl.trainer.ppo.core_algos import agg_loss, compute_policy_loss, get_policy_loss_fn, kl_penalty
+from verl.utils.dataset.vision_utils import process_multi_modal_inputs_for_minicpmo
 from verl.utils.debug import GPUMemoryLogger
 from verl.utils.device import get_device_id, get_device_name, is_cuda_available, is_npu_available
 from verl.utils.fsdp_utils import FSDPModule, fsdp2_clip_grad_norm_
@@ -88,13 +89,11 @@ class DataParallelPPOActor(BasePPOActor):
         response_length = micro_batch["responses"].size(-1)
         multi_modal_inputs = {}
         if "multi_modal_inputs" in micro_batch.keys():
-            for key in micro_batch["multi_modal_inputs"][0].keys():
-                # Special handling for MiniCPM-o model: pixel_values, image_bound, and tgt_sizes
-                # need different concatenation strategies compared to other multimodal inputs
-                if (key == "pixel_values" and isinstance(micro_batch["multi_modal_inputs"][0]["pixel_values"], list)) or key == "image_bound" or key == "tgt_sizes":
-                    # For MiniCPM-o: keep as list structure instead of concatenating tensors
+            if "image_bound" in micro_batch["multi_modal_inputs"][0]:  # minicpm-o logic
+                for key in micro_batch["multi_modal_inputs"][0].keys():
                     multi_modal_inputs[key] = [inputs[key] for inputs in micro_batch["multi_modal_inputs"]]
-                else:
+            else:
+                for key in micro_batch["multi_modal_inputs"][0].keys():
                     multi_modal_inputs[key] = torch.cat([inputs[key] for inputs in micro_batch["multi_modal_inputs"]], dim=0)
 
         with torch.autocast(device_type=self.device_name, dtype=torch.bfloat16):
@@ -116,25 +115,8 @@ class DataParallelPPOActor(BasePPOActor):
                 else:
                     position_ids_rmpad = index_first_axis(rearrange(position_ids.unsqueeze(-1), "b s ... -> (b s) ..."), indices).transpose(0, 1)
 
-                if "multi_modal_inputs" in micro_batch:
-                    # MiniCPM-o specific processing for image bounds and pixel values
-                    if "image_bound" in multi_modal_inputs:
-                        # Adjust image bounds based on left padding and cumulative sequence lengths
-                        # This is necessary for MiniCPM-o's vision-language alignment
-                        left_padding_length = torch.argmax(attention_mask, dim=1)
-                        image_bounds = []
-                        for i in range(len(multi_modal_inputs["image_bound"])):
-                            image_bound = multi_modal_inputs["image_bound"][i].to(left_padding_length.device) - left_padding_length[i] + cu_seqlens[i]
-                            image_bounds.append(image_bound)
-                        multi_modal_inputs["image_bound"] = [torch.vstack(image_bounds)]
-                        # Flatten pixel values list for MiniCPM-o processing
-                        pixel_values = []
-                        for i in range(len(multi_modal_inputs["pixel_values"])):
-                            pixel_values.extend([p for p in multi_modal_inputs["pixel_values"][i]])
-                        multi_modal_inputs["pixel_values"] = [pixel_values]
-                    # Handle target sizes for MiniCPM-o vision processing
-                    if "tgt_sizes" in multi_modal_inputs:
-                        multi_modal_inputs["tgt_sizes"] = [torch.vstack(multi_modal_inputs["tgt_sizes"])]
+                if "image_bound" in multi_modal_inputs:
+                    multi_modal_inputs = process_multi_modal_inputs_for_minicpmo(input_ids, attention_mask, position_ids, cu_seqlens, multi_modal_inputs)
 
                 # for compute the log_prob
                 input_ids_rmpad_rolled = torch.roll(input_ids_rmpad, shifts=-1, dims=1)  # (1, total_nnz)
